chore: rename article to feed/post.
