+++ src/components/FeedList/index.tsx
@@ -3,27 +3,27 @@ import { FC, forwardRef, useImperativeHandle } from "react";
 import { NativeScrollEvent, NativeSyntheticEvent } from "react-native";
 import Animated, { useAnimatedReaction, useAnimatedScrollHandler, useSharedValue } from "react-native-reanimated";
 import { Stack } from "tamagui";
-import { ArticleListItem, Props as ArticleListItemProps } from "./ArticleListItem";
+import { FeedListItem, Props as FeedListItemProps } from "./FeedListItem";
 
 export interface Props {
     onScroll?: ReturnType<typeof useAnimatedScrollHandler>
     onScrollEndDrag?: (event: NativeSyntheticEvent<NativeScrollEvent>) => void
 }
 
-export interface ArticleListInstance { }
+export interface FeedListInstance { }
 
-const data = Array.from({ length: 20 }).map((_, index) => ({ title: `Article ${index}` }))
+const data = Array.from({ length: 20 }).map((_, index) => ({ title: `Feed ${index}` }))
 
-export const ArticleList = forwardRef<ArticleListInstance, Props>((props, ref) => {
+export const FeedList = forwardRef<FeedListInstance, Props>((props, ref) => {
     const onScrollHandler = useComposedScrollHandler([props.onScroll]);
 
-    return <Animated.FlatList<ArticleListItemProps>
+    return <Animated.FlatList<FeedListItemProps>
         // TODO
         data={data}
         contentContainerStyle={{ padding: 16 }}
         renderItem={({ item, index }) => {
             return <Stack key={index} marginBottom={'$3'} >
-                <ArticleListItem {...item} />
+                <FeedListItem {...item} />
             </Stack>
         }}
         windowSize={3}
