+++ src/i18n/en.ts
@@ -1,7 +1,7 @@
 import zh_cn from "./zh-cn";
 
 const en: typeof zh_cn = {
-    article_list: "Article List",
+    feed: "Feed",
     latest: "Latest",
     popular: "Popular",
     follow: "Follow",
