+++ src/i18n/zh-cn.ts
@@ -1,5 +1,5 @@
 const zh_cn = {
-    article_list: '文章列表',
+    feed: '文章列表',
     latest: '最新',
     popular: '最热',
     follow: '关注',
